Remove Clock Pin requirement for hardware SPI
For the hardware SPI the clock PIN was required for the reset function even for the hardware SPI. An update implemented a reset function which does use a empty hardware spi push as replacement for the hardware spi. The software spi is handled seperately unchanged. Thus the requirements in the definition are changed here.